+++ database/migrations/2023_07_01_463540_create_saves_table.php
@@ -1,6 +1,5 @@
 <?php
 
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -9,28 +8,20 @@ return new class extends Migration
 {
     public function up()
     {
-        Schema::create('orders', function (Blueprint $table) {
+        if (Schema::hasTable('saves')) return;
+        Schema::create('saves', function (Blueprint $table) {
             $table->uuid('id')->primary();
             $table->foreignuuid('user_id');
             $table->foreignuuid('book_id');
-            $table->string('invoice_no')->unique();
-            $table->decimal('amount', 15, 5);
-            $table->decimal('total_amount', 15, 5);
-            $table->enum('status', ['processing', 'on_the_way', 'arrived', 'delivered', 'cancelled'])->default('processing');
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->foreign('book_id')->references('id')->on('books')->onDelete('cascade');
-
-
-
-            $table->index(['invoice_no']);
-            $table->index(['status']);
         });
     }
 
     public function down()
     {
-        Schema::dropIfExists('orders');
+        Schema::dropIfExists('saves');
     }
 };
